+++ unikmer/cmd/head.go
@@ -1,0 +1,178 @@
+// Copyright © 2018-2020 Wei Shen <shenwei356@gmail.com>
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+package cmd
+
+import (
+	"bufio"
+	"fmt"
+	"io"
+	"os"
+	"runtime"
+
+	"github.com/shenwei356/unikmer"
+	"github.com/spf13/cobra"
+)
+
+// headCmd represents
+var headCmd = &cobra.Command{
+	Use:   "head",
+	Short: "Extract the first N k-mers",
+	Long: `Extract the first N k-mers
+
+Attentions:
+  1. The 'canonical' flags of all files should be consistent.
+  2. Input files should ALL have or don't have taxid information.
+
+`,
+	Run: func(cmd *cobra.Command, args []string) {
+		opt := getOptions(cmd)
+		runtime.GOMAXPROCS(opt.NumCPUs)
+
+		var err error
+
+		if opt.Verbose {
+			log.Info("checking input files ...")
+		}
+		files := getFileListFromArgsAndFile(cmd, args, true, "infile-list", true)
+		if opt.Verbose {
+			if len(files) == 1 && isStdin(files[0]) {
+				log.Info("no files given, reading from stdin")
+			} else {
+				log.Infof("%d input file(s) given", len(files))
+			}
+		}
+
+		checkFileSuffix(extDataFile, files...)
+
+		outFile := getFlagString(cmd, "out-prefix")
+		N := getFlagNonNegativeInt(cmd, "number")
+
+		if !isStdout(outFile) {
+			outFile += extDataFile
+		}
+		outfh, gw, w, err := outStream(outFile, opt.Compress, opt.CompressionLevel)
+		checkError(err)
+		defer func() {
+			outfh.Flush()
+			if gw != nil {
+				gw.Close()
+			}
+			w.Close()
+		}()
+
+		var writer *unikmer.Writer
+
+		var infh *bufio.Reader
+		var r *os.File
+		var reader *unikmer.Reader
+		var code uint64
+		var taxid uint32
+		var k int = -1
+		var canonical bool
+		var hasTaxid bool
+		var firstFile = true
+		var n int
+		var flag int
+		var nfiles = len(files)
+
+		for i, file := range files {
+			if !firstFile && file == files[0] {
+				continue
+			}
+
+			if opt.Verbose {
+				log.Infof("processing file (%d/%d): %s", i+1, nfiles, file)
+			}
+
+			flag = func() int {
+				infh, r, _, err = inStream(file)
+				checkError(err)
+				defer r.Close()
+
+				reader, err = unikmer.NewReader(infh)
+				checkError(err)
+
+				if k == -1 {
+					k = reader.K
+					canonical = reader.IsCanonical()
+					hasTaxid = !opt.IgnoreTaxid && reader.HasTaxidInfo()
+
+					mode := reader.Flag
+					if hasTaxid {
+						mode |= unikmer.UNIK_INCLUDETAXID
+					}
+					writer, err = unikmer.NewWriter(outfh, k, mode)
+					checkError(err)
+					writer.SetMaxTaxid(maxUint32N(reader.GetTaxidBytesLength())) // follow reader
+				} else {
+					if k != reader.K {
+						checkError(fmt.Errorf("K (%d) of binary file '%s' not equal to previous K (%d)", reader.K, file, k))
+					}
+					if reader.IsCanonical() != canonical {
+						checkError(fmt.Errorf(`'canonical' flags not consistent, please check with "unikmer stats"`))
+					}
+					if !opt.IgnoreTaxid && reader.HasTaxidInfo() != hasTaxid {
+						if reader.HasTaxidInfo() {
+							checkError(fmt.Errorf(`taxid information not found in previous files, but found in this: %s`, file))
+						} else {
+							checkError(fmt.Errorf(`taxid information found in previous files, but missing in this: %s`, file))
+						}
+					}
+				}
+
+				for {
+					code, taxid, err = reader.ReadCodeWithTaxid()
+					if err != nil {
+						if err == io.EOF {
+							break
+						}
+						checkError(err)
+					}
+					if n >= N {
+						return flagBreak
+					}
+					n++
+					writer.WriteCodeWithTaxid(code, taxid)
+				}
+
+				return flagContinue
+			}()
+
+			if flag == flagReturn {
+				return
+			} else if flag == flagBreak {
+				break
+			}
+		}
+
+		checkError(writer.Flush())
+		if opt.Verbose {
+			log.Infof("%d k-mers saved to %s", n, outFile)
+		}
+	},
+}
+
+func init() {
+	RootCmd.AddCommand(headCmd)
+
+	headCmd.Flags().StringP("out-prefix", "o", "-", `out file prefix ("-" for stdout)`)
+	headCmd.Flags().IntP("number", "n", 10, "number of k-mers to extract")
+}